ME55: Resolve linter errors in PeriphDrivers/Include.

+++ MAX/Libraries/PeriphDrivers/Include/MAX32650/tmr.h
@@ -44,6 +44,7 @@
 #define _TMR_H_
 
 /* **** Includes **** */
+#include <stdint.h>
 #include "tmr_regs.h"
 #include "mxc_sys.h"
 #include "mxc_errors.h"
@@ -209,7 +210,7 @@ void MXC_TMR_SetCount(mxc_tmr_regs_t *tmr, uint32_t cnt);
  * @param      tmr   The timer
  * @param[in]  us    microseconds to delay for
  */
-void MXC_TMR_Delay(mxc_tmr_regs_t *tmr, unsigned long us);
+void MXC_TMR_Delay(mxc_tmr_regs_t *tmr, uint32_t us);
 
 /**
  * @brief      Start a timer that will time out after a certain number of microseconds
@@ -217,7 +218,7 @@ void MXC_TMR_Delay(mxc_tmr_regs_t *tmr, unsigned long us);
  * @param      tmr   The timer
  * @param[in]  us    microseconds to time out after
  */
-void MXC_TMR_TO_Start(mxc_tmr_regs_t *tmr, unsigned long us);
+void MXC_TMR_TO_Start(mxc_tmr_regs_t *tmr, uint32_t us);
 
 /**
  * @brief      Check on time out timer
